Moves AztecTextAttachment to its own file
